requirements.txt
fastapi
uvicorn
tensorflow
python-multipart
pillow
numpy
